requirements.txt
click==7.1.2
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
python-dotenv==1.0.1
scikit_learn==1.3.0
setuptools==68.2.2
torch==2.2.1
torchvision==0.17.1
